Fix coverity warnings from static analysis tools
1. TAINTED_SCALAR
   a. phNxpEseProto7816_3.cpp: phNxpEseProto7816_DecodeFrame()
      Check for valid data length
2. INTEGER_OVERFLOW
   a. phNxpEse_Api.cpp: phNxpEse_read()
      Read packet length can not be negative

+++ libese-spi/p73/lib/phNxpEse_Api.cpp
@@ -1253,7 +1253,7 @@ ESESTATUS phNxpEse_read(uint32_t* data_len, uint8_t** pp_data) {
     *pp_data = nxpese_ctxt.p_read_buff;
     status = ESESTATUS_FAILED;
   } else {
-    PH_PAL_ESE_PRINT_PACKET_RX(nxpese_ctxt.p_read_buff, ret);
+    PH_PAL_ESE_PRINT_PACKET_RX(nxpese_ctxt.p_read_buff, (uint16_t)ret);
     *data_len = ret;
     *pp_data = nxpese_ctxt.p_read_buff;
     status = ESESTATUS_SUCCESS;
